refactor(services): replace deep imports with public API imports (#2442)
