+++ src/vibe3/services/branch_arg.py
@@ -1,10 +1,7 @@
 """Backward compatibility — migrated to services.shared.branches."""
 
 # Re-export imports for test patch compatibility
-from vibe3.config.convention_resolver import (  # noqa: F401
-    ConventionResolver,
-    get_convention,
-)
+from vibe3.config import ConventionResolver, get_convention  # noqa: F401
 from vibe3.services.flow_service import FlowService  # noqa: F401
 from vibe3.services.shared.branches import (  # noqa: F401
     resolve_branch_and_issue,
