+++ src/vibe3/services/check_pr_service.py
@@ -522,10 +522,10 @@ The unresolved work continues in #{bridge_issue_number}.
             )
 
         # Check if issue already closed
-        from vibe3.clients.github_field_constants import GITHUB_FIELDS_STATE_ONLY
+        from vibe3.clients import GITHUB_FIELDS_STATE_ONLY
 
         gh_issue = self.github_client.view_issue(
-            task_issue_number, fields=list(GITHUB_FIELDS_STATE_ONLY)
+            task_issue_number, fields=list(GITHUB_FIELDS_STATE_ONLY)  # type: ignore[call-overload]
         )
         if not isinstance(gh_issue, dict):
             # Failed to fetch issue data (network error or not found)
