+++ src/vibe3/services/check_service.py
@@ -318,13 +318,11 @@ class CheckService(CheckRemote):
             orchestration_state: IssueState | None = None
             issue_payload: dict | None = None
             if task_issue:
-                from vibe3.clients.github_field_constants import (
-                    GITHUB_DEFAULT_VIEW_FIELDS,
-                )
+                from vibe3.clients import GITHUB_DEFAULT_VIEW_FIELDS
 
                 # Need state, labels, and body for state validation
                 issue = self.github_client.view_issue(
-                    task_issue, fields=list(GITHUB_DEFAULT_VIEW_FIELDS)
+                    task_issue, fields=list(GITHUB_DEFAULT_VIEW_FIELDS)  # type: ignore[call-overload]
                 )
                 if issue == "network_error":
                     issues.append(
