+++ src/vibe3/services/comment_service.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 import re
 from typing import Any
 
+from vibe3.config import get_manager_usernames as _get_manager_usernames
 from vibe3.config import load_orchestra_config
-from vibe3.config.manager_config import get_manager_usernames as _get_manager_usernames
 from vibe3.utils.constants import AUTOMATED_MARKERS, GENERIC_AGENT_MARKER_PATTERN
