+++ src/vibe3/services/convention_resolver.py
@@ -1,7 +1,3 @@
-from vibe3.config.convention_resolver import (
-    ConventionResolver,
-    get_convention,
-    get_resolver,
-)
+from vibe3.config import ConventionResolver, get_convention, get_resolver
 
 __all__ = ["ConventionResolver", "get_convention", "get_resolver"]
