+++ src/vibe3/services/issue/context.py
@@ -25,10 +25,10 @@ def load_issue_info(
         UserError: If issue cannot be loaded or parsed
     """
     github = github or GitHubClient()
-    from vibe3.clients.github_field_constants import GITHUB_DEFAULT_VIEW_FIELDS
+    from vibe3.clients import GITHUB_DEFAULT_VIEW_FIELDS
 
     payload = github.view_issue(
-        issue_number, repo=config.repo, fields=list(GITHUB_DEFAULT_VIEW_FIELDS)
+        issue_number, repo=config.repo, fields=list(GITHUB_DEFAULT_VIEW_FIELDS)  # type: ignore[call-overload]
     )
 
     if payload == "network_error":
